# openapi.yaml
openapi: 3.0.3
info:
  title: 由岐商城 API
  version: 1.0.0
  description: 基于 yuki-shopping 数据库设计的京东类商城学习项目接口。
servers:
  - url: /api/v1
    description: 当前服务
tags:
  - name: 认证
  - name: 用户地址
  - name: 商品
  - name: 购物车
  - name: 订单支付
  - name: 评价
  - name: 管理端
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [认证]
      security: []
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200': {$ref: '#/components/responses/Success'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /auth/login:
    post:
      tags: [认证]
      security: []
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiResponse'}
  /auth/refresh:
    post:
      tags: [认证]
      security: []
      summary: 刷新令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: {refreshToken: {type: string}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /users/me:
    get:
      tags: [用户地址]
      summary: 当前用户资料
      responses: {'200': {$ref: '#/components/responses/Success'}}
    put:
      tags: [用户地址]
      summary: 修改当前用户资料
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: {type: string, maxLength: 50}
                phone: {type: string}
                email: {type: string, format: email}
                avatar: {type: string, format: uri}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /users/me/addresses:
    get:
      tags: [用户地址]
      summary: 地址列表
      responses: {'200': {$ref: '#/components/responses/Success'}}
    post:
      tags: [用户地址]
      summary: 新增地址
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressRequest'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /users/me/addresses/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
    put:
      tags: [用户地址]
      summary: 修改地址
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AddressRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    delete:
      tags: [用户地址]
      summary: 删除地址
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /users/me/addresses/{id}/default:
    put:
      tags: [用户地址]
      summary: 设置默认地址
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /categories/tree:
    get:
      tags: [商品]
      security: []
      summary: 分类树
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /brands:
    get:
      tags: [商品]
      security: []
      summary: 品牌分页
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /products:
    get:
      tags: [商品]
      security: []
      summary: 商品分页查询
      parameters:
        - {name: keyword, in: query, schema: {type: string}}
        - {name: categoryId, in: query, schema: {type: integer, format: int64}}
        - {name: brandId, in: query, schema: {type: integer, format: int64}}
        - {name: minPrice, in: query, schema: {type: string}}
        - {name: maxPrice, in: query, schema: {type: string}}
        - {name: sort, in: query, schema: {type: string, enum: [default, price_asc, price_desc, sales, newest]}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /products/{productId}:
    get:
      tags: [商品]
      security: []
      summary: 商品详情
      parameters: [{name: productId, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /products/{productId}/reviews:
    get:
      tags: [评价]
      security: []
      summary: 商品评价列表
      parameters:
        - {name: productId, in: path, required: true, schema: {type: integer, format: int64}}
        - {name: rating, in: query, schema: {type: integer, minimum: 1, maximum: 5}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /cart:
    get:
      tags: [购物车]
      summary: 购物车列表
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /cart/items:
    post:
      tags: [购物车]
      summary: 加入购物车
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CartItemRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /cart/items/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
    put:
      tags: [购物车]
      summary: 修改购物车条目
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {quantity: {type: integer, minimum: 1}, checked: {type: integer, enum: [0, 1]}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    delete:
      tags: [购物车]
      summary: 删除购物车条目
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/preview:
    post:
      tags: [订单支付]
      summary: 订单预览
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrderCreateRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders:
    get:
      tags: [订单支付]
      summary: 我的订单
      parameters:
        - {name: status, in: query, schema: {type: integer, enum: [10, 20, 30, 40, 50, 60]}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    post:
      tags: [订单支付]
      summary: 创建订单
      parameters: [{name: Idempotency-Key, in: header, required: true, schema: {type: string}}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrderCreateRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/{orderNo}:
    get:
      tags: [订单支付]
      summary: 订单详情
      parameters: [{name: orderNo, in: path, required: true, schema: {type: string}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/{orderNo}/cancel:
    post:
      tags: [订单支付]
      summary: 取消订单
      parameters: [{name: orderNo, in: path, required: true, schema: {type: string}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/{orderNo}/confirm:
    post:
      tags: [订单支付]
      summary: 确认收货
      parameters: [{name: orderNo, in: path, required: true, schema: {type: string}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/{orderNo}/pay:
    post:
      tags: [订单支付]
      summary: 发起支付
      parameters:
        - {name: orderNo, in: path, required: true, schema: {type: string}}
        - {name: Idempotency-Key, in: header, schema: {type: string}}
      requestBody:
        content: {application/json: {schema: {type: object, properties: {payType: {type: integer, enum: [1, 2, 3], default: 1}}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /payments/callback:
    post:
      tags: [订单支付]
      security: []
      summary: 支付回调
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [paymentNo, amount, status], properties: {paymentNo: {type: string}, amount: {type: string}, status: {type: integer}, tradeNo: {type: string}}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /orders/{orderNo}/items/{itemId}/review:
    post:
      tags: [评价]
      summary: 发表评价
      parameters:
        - {name: orderNo, in: path, required: true, schema: {type: string}}
        - {name: itemId, in: path, required: true, schema: {type: integer, format: int64}}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ReviewRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /admin/auth/login:
    post:
      tags: [管理端]
      security: []
      summary: 管理员登录
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /admin/orders/{orderNo}/ship:
    post:
      tags: [管理端]
      summary: 订单发货
      parameters: [{name: orderNo, in: path, required: true, schema: {type: string}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expressCompany, expressNo]
              properties: {expressCompany: {type: string}, expressNo: {type: string}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /admin/reviews/{id}/status:
    put:
      tags: [管理端]
      summary: 显示或隐藏评价
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: integer, enum: [0, 1]}}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  responses:
    Success:
      description: 请求成功
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    BadRequest:
      description: 参数错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
  schemas:
    ApiResponse:
      type: object
      properties:
        code: {type: integer, example: 0}
        message: {type: string, example: ok}
        data: {nullable: true}
        traceId: {type: string}
    RegisterRequest:
      type: object
      required: [username, password, nickname]
      properties:
        username: {type: string, minLength: 3, maxLength: 50}
        password: {type: string, minLength: 8, writeOnly: true}
        nickname: {type: string, maxLength: 50}
        phone: {type: string}
        email: {type: string, format: email}
    LoginRequest:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string, writeOnly: true}}
    AddressRequest:
      type: object
      required: [receiverName, receiverPhone, province, city, district, detail]
      properties:
        receiverName: {type: string}
        receiverPhone: {type: string}
        province: {type: string}
        city: {type: string}
        district: {type: string}
        detail: {type: string}
        isDefault: {type: integer, enum: [0, 1], default: 0}
    CartItemRequest:
      type: object
      required: [skuId, quantity]
      properties:
        skuId: {type: integer, format: int64}
        quantity: {type: integer, minimum: 1}
    OrderCreateRequest:
      type: object
      required: [addressId, items]
      properties:
        addressId: {type: integer, format: int64}
        items:
          type: array
          minItems: 1
          items: {type: object, required: [skuId, quantity], properties: {skuId: {type: integer, format: int64}, quantity: {type: integer, minimum: 1}}}
        remark: {type: string, maxLength: 200}
    ReviewRequest:
      type: object
      required: [rating]
      properties:
        rating: {type: integer, minimum: 1, maximum: 5}
        content: {type: string, maxLength: 1000}
        images: {type: array, items: {type: string, format: uri}}
        isAnonymous: {type: integer, enum: [0, 1], default: 0}